BlenderKit: resolutions
This introduces resolutins into the addon.
This update should enhance the usability of the addon, especially for people with weaker computers.
It downloads reduced version of the assets - only images are scaled down by now. Images are also converted in some cases from .png to .jpgs to save space.

 - there's a default resolution setting
 - resolutions can be swapped by user
 - resolutions apply only to textured models and materials with textures larger than 1024px
 - Resolutions aren't yet generated on the server, so will be visible after a few days.

Version of the addon was bumped up to 1.0.40.