Past ratings are now loaded for all logged in users
the ratings are loaded when popup (right click or fast rate) are called. This causes a delay.

The delay is relatively big - several seconds or until the user doesn't mouse over the element.

This commit also reduces task_queue step - from 2 seconds to 0.3 seconds. This shouldn't impact performance if the queue is empty - the funciton now returns right at the beginning.