Force category when uploading assets
Ui is now red and there are now 'None' categories, which are default.
This forces setting category and avoids assets to accumulate in default architecture - building - commercial category.
Disadvantage is that the category setting on existing assets inside .blend gets changed(but that would change also if category is added or any changes in category system, since only index is stored)