Fix upload in case not all textures are packed correctly.
This is a blender issue where blender sometimes wants to pack textures that aren't actually needed and are somehow still in the project.
This prevented upload completely, where the task of the creator is to check correct upload anyway.
Not sure how to pass a warning to the user, since it's all in a bg process that is run from daemon.