Fix an issue where  you:
swithc to scenes,
search something
creaeta a new file - models are active, but old results are there
try to drag drop something, it fails.

Why to fix it this way:
This prepares more ground for mixing asset types that get dragged into scene, having this as a scenewide setting is weird.