Fix #730
This solves this hard issue in a quite complex way, hopefully expandable/improvable in the future

now it tries to evaluate the mesh and map the materials to material slots + set material nodes in gn tree.
Creates a mapping that is later used on the object when material finishes downloading.
CLicking in assetbar sometimes assigns GN set material slot instead, if it's first
anyway, slots mixed with GN will always be a mess and let's assume the user can assign himself if this doesn't work, so it's not a 100% solution
Maybe blender's asset manager has solved this? not sure...