Skip saving preferences on factory startup
When running Blender with factory default settings, via `--factory-startup`, factory default preferences are used and user preferences are skipped and not saved (autosave is disabled). However, trying to enable this add-on (which fails, but can be a separate fix) triggers an unexpected save to user preferences, causing the factory startup settings to override the user's. This prevents that.