Try to fix #522 - drawing and timers interference
This attempt on fixing tests the context before doing any of the operations that involve changing blender data, and adding random offsets to timers (which strangely seems to work a bit better). Still, there are errors that clearly stem from bad internal timing of blender's processes, because any operator in a chain of operators can cause the error even after context was already tested and accessible in the same function run.